backend/requirements.txt
langchain
openai
tiktoken
faiss-cpu
pypdf
unstructured
python-docx
fastapi
uvicorn
pydantic
python-dotenv
aiofiles
